best effort training, plus a few tiny cleanups
components:
be_training/ - contains the cpp extensions for training a background resnet50 job.
This includes a patch for pytorch, and a script to build/install pytorch from source with the patch.
To install the cpp extension, run python setup.py install in be_training/. The python interface for
the extension is in torch_be_training.cpp.

JobContext idle time tracking: instrument ReceiveSamplesFunc/SendSamplesFunc to report idle times to the JobContext.
These times are profiled across iterations 50-100. The tracking
is similar to timetrace, but totally separate. After profiling, JobContext reports idle times to Runtime.